[✨feat] 전역 상태 관리를 위해 상태 관리 라이브러리 Recoil 추가(VOTEGUARD-25)
- 원래는 부모 컴포넌트에서 자식 컴포넌트에 props로 데이터를 전달하는 형태로 구현한 상태
- 하지만 컴포넌트의 깊이가 깊어지면서 불편함을 느끼게 되었고 Recoil을 추가해서 전역 상태 관리를 할 수 있게 변경함
