+++ Front/vote-guardian/package.json
@@ -15,6 +15,7 @@
     "react-dom": "^18",
     "react-icons": "^5.0.1",
     "react-responsive": "^9.0.2",
+    "recoil": "^0.7.7",
     "sass": "^1.70.0"
   },
   "devDependencies": {
