+++ Front/vote-guardian/src/recoil/atoms/createVoteListState.js
@@ -1,0 +1,7 @@
+const { atom } = require("recoil");
+
+export const createVoteListState = atom({
+    //key : 전역적인 고유 키, default : 초기 값
+    key : 'createVoteListState',
+    default: []
+})
